requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy>=2.0.25
greenlet>=3.0.0
alembic==1.13.1
psycopg-binary
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
google-auth==2.26.2
httpx==0.26.0
pydantic-settings==2.1.0
websockets==12.0
requests==2.31.0
fastapi-mail
